requirements.txt
anthropic>=0.40.0
streamlit>=1.40
python-dotenv>=1.0
python-frontmatter>=1.1
jsonschema>=4.23
pyyaml>=6.0
pandas>=2.2
